rename directory for content page components
