+++ bikespace_frontend/src/app/(landing)/page.mdx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import {HeroBlock} from '@/components/landing/content-blocks';
+import {HeroBlock} from '@/components/content-pages/content-blocks';
 
 import bikespaceIntro from '@/assets/images/Hero_Concept_v2_web.jpg';
