+++ bikespace_frontend/src/app/about/page.mdx
@@ -4,7 +4,7 @@ import {
   FeatureBoxWrapper,
   FeatureBox,
   IconHeading,
-} from '@/components/landing/content-blocks';
+} from '@/components/content-pages/content-blocks';
 
 import featureBoxParkingMap from '@/assets/images/parking-map-mobile.png';
 import featureBoxApp from '@/assets/images/mobile_reporting_app.png';
