+++ bikespace_frontend/src/components/content-pages/content-blocks/content-blocks.module.scss
