+++ bikespace_frontend/src/components/content-pages/content-page-layout/index.ts
