+++ bikespace_frontend/src/components/content-pages/footer/Footer.tsx
