+++ bikespace_frontend/src/components/content-pages/footer/index.ts
