+++ bikespace_frontend/src/components/content-pages/index.ts
@@ -1,0 +1,1 @@
+export {ContentPageLayout} from './content-page-layout';
