--- bikespace_frontend/src/components/landing/index.ts
@@ -1,1 +1,0 @@
-export {ContentPageLayout} from './landing-layout';
